Repository and Controller update

+++ LibraryDB/Books.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Books]
 (
-	[isbn] INT NOT NULL PRIMARY KEY,
+	[isbn] BIGINT NOT NULL PRIMARY KEY,
     [title] VARCHAR(255) NOT NULL,
 	[authors] VARCHAR(255) NOT NULL,
 	[publisher] VARCHAR(255) NOT NULL,
